Move LoginFormAuthenticator directly in SiteConfig namespace
